pyproject.toml
[project]
name = "batman"
version = "0.1.0"
description = "Local Video Auto-Label → Human Correct → Fine-Tune Real-Time Detector"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.12",
    "websockets>=13.0",
    
    # Database
    "sqlalchemy>=2.0.35",
    "aiosqlite>=0.20.0",
    
    # ML/Vision
    "torch>=2.4.0",
    "torchvision>=0.19.0",
    "ultralytics>=8.3.237",  # SAM3 support
    "supervision>=0.25.0",
    "timm>=1.0.0",  # Required by SAM3
    "clip @ git+https://github.com/ultralytics/CLIP.git",  # Required by SAM3 text prompts
    "rfdetr[metrics]>=1.0.0",  # RF-DETR with TensorBoard support
    "tensorboard>=2.15.0",  # Training visualization
    
    # Video/Image processing
    "opencv-python>=4.10.0",
    "ffmpeg-python>=0.2.0",
    "pillow>=10.4.0",
    "numpy>=1.26.0",
    "scipy>=1.13.0",
    
    # Utilities
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "aiofiles>=24.1.0",
    "httpx>=0.27.0",
    "tqdm>=4.66.0",
    "loguru>=0.7.0",
    "pyyaml>=6.0.0",
    
    # Development
    "ipykernel>=7.1.0",
]

[project.scripts]
batman = "backend.app.main:run_server"

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.6.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
